delete car functionality corrected

+++ src/main/java/com/udacity/vehicles/service/CarService.java
@@ -140,19 +140,22 @@ public class CarService {
          * TODO: Find the car by ID from the `repository` if it exists.
          *   If it does not exist, throw a CarNotFoundException
          */
-//        Optional<Car> optionalCar = repository.findById(id);
-//        if (optionalCar.isPresent()){
-//            car = optionalCar.get();
-//        }else {
-//            throw new CarNotFoundException();
-//        }
+        Optional<Car> optionalCar = repository.findById(id);
+        if (optionalCar.isPresent()){
+            car = optionalCar.get();
+            repository.delete(car);
+        }else {
+            throw new CarNotFoundException();
+        }
 
-        car = repository.findById(id).orElseThrow(() -> new CarNotFoundException());
-        /**
-         * TODO: Delete the car from the repository.
-         */
-        repository.delete(car);
-        //repository.deleteById(car.getId());
+
+
+//        car = repository.findById(id).orElseThrow(() -> new CarNotFoundException());
+//        /**
+//         * TODO: Delete the car from the repository.
+//         */
+//        repository.delete(car);
+//        //repository.deleteById(car.getId());
 
 
     }
